requirements.txt
camelot-py==1.0.0
pandas==2.3.1
numpy==2.3.1
fastapi==0.110.0
uvicorn==0.29.0
SQLAlchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.6
pydantic==2.11.7
pytest==8.4.1